rename global localization and removed linting on it
