+++ components/activity/type/lang/localization.js
@@ -1,4 +1,4 @@
-import { getLocalizeResources } from '../../../../lang/localization.js';
+import { getLocalizeResources } from '../../../../lang/localizeResources.js';
 import { LocalizeMixin } from '@brightspace-ui/core/mixins/localize-mixin.js';
 
 export const LocalizeFoundationType = superclass => class extends LocalizeMixin(superclass) {
